aapt2: refactor testsuite
 - split from aapt1/aapt2 (those that do building)
 - tests that just test decode (no aapt) split
 - categories for androlib/encdoers/util for various others
