+++ brut.apktool/apktool-lib/src/test/java/brut/androlib/aapt1/DebugTagRetainedTest.java
@@ -14,8 +14,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib;
+package brut.androlib.aapt1;
 
+import brut.androlib.*;
 import brut.directory.ExtFile;
 import brut.common.BrutException;
 import brut.util.OS;
@@ -29,27 +30,32 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 /**
  * @author Connor Tumbleson <connor.tumbleson@gmail.com>
  */
-public class MinifiedArscTest extends BaseTest {
+public class DebugTagRetainedTest extends BaseTest {
 
     @BeforeClass
     public static void beforeClass() throws Exception {
         TestUtils.cleanFrameworkFile();
         sTmpDir = new ExtFile(OS.createTempDirectory());
-        TestUtils.copyResourceDir(MinifiedArscTest.class, "brut/apktool/issue1157/", sTmpDir);
+        sTestOrigDir = new ExtFile(sTmpDir, "issue1235-orig");
+        sTestNewDir = new ExtFile(sTmpDir, "issue1235-new");
+        LOGGER.info("Unpacking issue1235...");
+        TestUtils.copyResourceDir(DebugTagRetainedTest.class, "brut/apktool/issue1235/", sTestOrigDir);
 
-        String apk = "issue1157.apk";
-        sTestNewDir = new ExtFile(sTmpDir, "issue1157");
+        LOGGER.info("Building issue1235.apk...");
+        ApkOptions apkOptions = new ApkOptions();
+        apkOptions.debugMode = true;
 
-        // decode issue1157.apk
-        ApkDecoder apkDecoder = new ApkDecoder(new ExtFile(sTmpDir, apk));
-        apkDecoder.setForceDelete(true);
+        File testApk = new File(sTmpDir, "issue1235.apk");
+        new Androlib(apkOptions).build(sTestOrigDir, testApk);
+
+        LOGGER.info("Decoding issue1235.apk...");
+        ApkDecoder apkDecoder = new ApkDecoder(testApk);
         apkDecoder.setOutDir(sTestNewDir);
-
-        // this should not raise an exception:
         apkDecoder.decode();
     }
 
@@ -59,14 +65,20 @@ public class MinifiedArscTest extends BaseTest {
     }
 
     @Test
-    public void checkIfMinifiedArscLayoutFileMatchesTest() throws IOException {
-        String expected = TestUtils.replaceNewlines("<?xml version=\"1.0\" encoding=\"utf-8\"?>\n" +
-                "<LinearLayout n1:orientation=\"vertical\" n1:layout_width=\"fill_parent\" n1:layout_height=\"fill_parent\"\n" +
-                "  xmlns:n1=\"http://schemas.android.com/apk/res/android\">\n" +
-                "    <com.ibotpeaches.issue1157.MyCustomView n1:max=\"100\" n2:default_value=\"1.0\" n2:max_value=\"5.0\" n2:min_value=\"0.2\" xmlns:n2=\"http://schemas.android.com/apk/res-auto\" />\n" +
-                "</LinearLayout>");
+    public void buildAndDecodeTest() {
+        assertTrue(sTestNewDir.isDirectory());
+    }
 
-        byte[] encoded = Files.readAllBytes(Paths.get(sTestNewDir + File.separator + "res" + File.separator + "xml" + File.separator + "custom.xml"));
+    @Test
+    public void DebugIsTruePriorToBeingFalseTest() throws IOException {
+        String apk = "issue1235-new";
+
+        String expected = TestUtils.replaceNewlines("<?xml version=\"1.0\" encoding=\"utf-8\" standalone=\"no\"?>\n" +
+                "<manifest xmlns:android=\"http://schemas.android.com/apk/res/android\" package=\"com.ibotpeaches.issue1235\" platformBuildVersionCode=\"23\" platformBuildVersionName=\"6.0-2438415\">\n" +
+                "    <application android:debuggable=\"true\"/>" +
+                "</manifest>");
+
+        byte[] encoded = Files.readAllBytes(Paths.get(sTmpDir + File.separator + apk + File.separator + "AndroidManifest.xml"));
         String obtained = TestUtils.replaceNewlines(new String(encoded));
         assertEquals(expected, obtained);
     }
