+++ brut.apktool/apktool-lib/src/test/java/brut/androlib/aapt1/ReferenceVersionCodeTest.java
@@ -14,23 +14,35 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib;
+package brut.androlib.aapt1;
 
+import brut.androlib.Androlib;
+import brut.androlib.ApkDecoder;
+import brut.androlib.BaseTest;
+import brut.androlib.TestUtils;
+import brut.androlib.meta.MetaInfo;
 import brut.directory.ExtFile;
 import brut.common.BrutException;
 import brut.util.OS;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
 import java.io.File;
 import java.io.IOException;
 
-import org.junit.*;
+import static org.junit.Assert.assertEquals;
 
-public class LargeIntsInManifestTest extends BaseTest {
+/**
+ * @author Connor Tumbleson <connor.tumbleson@gmail.com>
+ */
+public class ReferenceVersionCodeTest extends BaseTest {
 
     @BeforeClass
     public static void beforeClass() throws Exception {
         TestUtils.cleanFrameworkFile();
         sTmpDir = new ExtFile(OS.createTempDirectory());
-        TestUtils.copyResourceDir(LargeIntsInManifestTest.class, "brut/apktool/issue767/", sTmpDir);
+        TestUtils.copyResourceDir(ReferenceVersionCodeTest.class, "brut/apktool/issue1234/", sTmpDir);
     }
 
     @AfterClass
@@ -39,28 +51,16 @@ public class LargeIntsInManifestTest extends BaseTest {
     }
 
     @Test
-    public void checkIfLargeIntsAreHandledTest() throws BrutException, IOException {
-        String apk = "issue767.apk";
+    public void referenceBecomesLiteralTest() throws BrutException, IOException {
+        String apk = "issue1234.apk";
 
-        // decode issue767.apk
+        // decode issue1234.apk
         ApkDecoder apkDecoder = new ApkDecoder(new File(sTmpDir + File.separator + apk));
-        sTestOrigDir = new ExtFile(sTmpDir + File.separator + apk + ".out");
-
+        ExtFile decodedApk = new ExtFile(sTmpDir + File.separator + apk + ".out");
         apkDecoder.setOutDir(new File(sTmpDir + File.separator + apk + ".out"));
         apkDecoder.decode();
 
-        // build issue767
-        ExtFile testApk = new ExtFile(sTmpDir, apk + ".out");
-        new Androlib().build(testApk, null);
-        String newApk = apk + ".out" + File.separator + "dist" + File.separator + apk;
-
-        // decode issue767 again
-        apkDecoder = new ApkDecoder(new File(sTmpDir + File.separator + newApk));
-        sTestNewDir = new ExtFile(sTmpDir + File.separator + apk + ".out.two");
-
-        apkDecoder.setOutDir(new File(sTmpDir + File.separator + apk + ".out.two"));
-        apkDecoder.decode();
-
-        compareXmlFiles("AndroidManifest.xml");
+        MetaInfo metaInfo = new Androlib().readMetaFile(decodedApk);
+        assertEquals("v1.0.0", metaInfo.versionInfo.versionName);
     }
 }
